Reduce number of thumbnails created to get v2 site building again (#6004)
* Reduce number of thumbnails created to get v2 site building again

* Change resolutions to resize

* Query allSitesYaml in StaticQuery so every page doesn't include the same data

* Move guts of showcase-detail to seperate component as can't have regular page + staticquery in same file
