+++ www/src/pages/showcase.js
@@ -863,7 +863,7 @@ export default ShowcasePage
 
 export const showcaseQuery = graphql`
   query ShowcaseQuery {
-    featured: allSitesYaml(filter: { featured: { eq: true } }) {
+    featured: allSitesYaml(limit: 1, filter: { featured: { eq: true } }) {
       edges {
         node {
           id
@@ -887,7 +887,10 @@ export const showcaseQuery = graphql`
         }
       }
     }
-    allSitesYaml(filter: { main_url: { ne: null } }) {
+    allSitesYaml(
+      limit: 1
+      filter: { featured: { eq: true }, main_url: { ne: null } }
+    ) {
       edges {
         node {
           id
